+++ .changeset/steady-runtimes-split.md
@@ -1,0 +1,5 @@
+---
+"@shipfox/api-workflows": patch
+---
+
+Splits workflow runtime internals into scheduling and step-config modules without changing behavior.
