+++ libs/api/workflows/src/core/step-transition/decide-step-transition.ts
@@ -1,5 +1,5 @@
 import type {Step, StepStatus} from '../entities/step.js';
-import type {RuntimeCompletionStatus} from '../workflow-runtime/runtime-dag.js';
+import type {RuntimeCompletionStatus} from '../workflow-scheduling/runtime-dag.js';
 
 const TERMINAL_STATUSES: ReadonlySet<StepStatus> = new Set(['succeeded', 'failed', 'cancelled']);
